Add created custom node to database
